pyproject.toml
[project]
name = "stata-bot"
version = "0.1.0"
description = "Stata LoRA Fine-tuning"
requires-python = ">=3.10,<3.11"
dependencies = [
    "transformers",
    "peft",
    "datasets",
    "trl",
    "python-dotenv",
    "pyyaml",
    "hf-transfer>=0.1.9",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "bitsandbytes>=0.49.1",
    "unsloth>=2026.1.4",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu126", marker = "sys_platform == 'linux'" }
torchvision = { index = "pytorch-cu126", marker = "sys_platform == 'linux'" }
torchaudio = { index = "pytorch-cu126", marker = "sys_platform == 'linux'" }
